optionsroms: consider ATI/AMD PCI graphics option ROM as invalid
That option ROM has a bytecode which requires proper support by a parser
to parse that bytecode. Trying to run QEMU with PCI-passthrough with the
Radeon HD 5450 card hangs the machine due to trying to run the code from
the PCI option ROM as-is.
We can detect such option ROM with the special string in offset 0x30 of
the option ROM.